+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project uses [Semantic Versioning](https://semver.org/lang/en/).
 
+## [Unreleased]
+
+### Fixed
+
+-   Unit tests: use `path.resolve` in `vi.mock` to ensure stable CI runs across OS.
+
 ## [1.0.2] - 2025-09-24
 
 ### Added